+++ examples/resources/vapi_file/resource.tf
@@ -1,3 +1,3 @@
-resource "vapi_file" "test" {
+resource "vapi_file" "test-vapi_file" {
   file_path = "/tmp/file.txt"
 }