+++ internal/provider/vapi_twilio_phone_number_resource_test.go
@@ -1,0 +1,8 @@
+package provider
+
+import (
+	"testing"
+)
+
+func TestAccVAPIPhoneNumberResource(t *testing.T) {
+}