+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 0.2.0 (Unreleased)
+## 0.3.0
 
 FEATURES:
+  - added VAPI Twilio Phone Number import - no fallback implemented
+  - fixed VAPI files support - issue when target file manually deleted from the VAPI service
+
+## 0.2.0
+
+FEATURES:
+  - added VAPI files support -- buggy as hell